Add Appeal#certified? method
In particular the certification_date being set means this case has
already been advanced. However, this will return *true* for remands,
since the certification date won't get re-set when re-certified to the
board.

Remands aren't covered in the current codebase, it's likely we'll need
a `remand?` method to handle that.