Rename `ready_for_appeal?` to `documents_match?`
In reality, the ready_for_appeal method isn't a construct that we use
in the code, we really want to know if the documents match. It's not
clear what this method should do in cases where the documents match but
the case isn't ready for certification (e.g. it has already been
certified), so this method has been renmaed to make that more clear.