Add in a Help page. Close: #53
Initial cut at a Help page, just start to put some basic things out in
the codebase, so that we can iterate on this later.

This is a straight HTML port of the Document Kavi wrote up. Needs more
work. This is just to get it into VCS.

Thanks, @KHarshawat