add other ROs used in VACOLS
Jed sent me the complete list of all ROs he uses in VACOLS. I manually compared that list of ROs with the list being used by caseflow. This patch is to add the missing ROs to caseflow.

Note: new ROs get added as business needs change. Jed has added several this year already, so we need to be aware that this is a dynamic, changing list. Jed suggested that the "proper" longer term fix would be to have caseflow dynamically get the latest list of ROs from VACOLS at runtime. There are details to figure out, like caching results for runtime speed, how to handle VACOLS being unavailable, etc. However, for now, it is still good to at least make this list match the latest of VACOLS, hence this patch.