Give routable special issues priority (#954)
* Fix some date issues with the test

Two tests will always fail from 7pm-12am EST because the date is in UAT
but the current user time zone is EST.

Also Timecop needs to apply to the test data too.

* If a claim has a routable special issue, don't let unsupported issues get

in their way.

connects #911

* Fix issue with date transforming to browser time zone.

This timezone issue was more gnarly than I thought.

`Date.new(timestring)` converts the `timestring` to the time zone local
to the browser and there is no way to change that without using moment.js.

This can cause an off-by-one error with the date if the timezone difference
between the RO and the browser timezone is larger than the time since the last
day or until the next day. This is particularly annoying in tests.

Instead of introducting moment.js right now, just pass the date (which
is what matters) without time and zone information. This guarantees that
the date is not switched because of weird timezone issues.

* Appease the linting gods

* Deal with small typo