Scope new Establish Claim Task creation to decision type
Every Establish Claim Task we tried to create today had multiple
appeals for the file number, which is annoying. The way to fix this
is to scope it to the type of decision.

Also, added refactored a little bit and added some better error
messaging.

No tests because this is a temporary feature.