Add cached values to EstablishClaim to improve performance
This reduces the VACOLS calls on the case worker task view, and
the admin create claim view by a lot. We could also implement this
on the missing decsions page, but there are some questions on whether
decision_date can change.

Caching can be dangerous, so as a precaution i have:
- Named cached values explicitly (ie: `cached_decision_type`), this
  way it's hard to fall into the trap of unknowingly pulling a
  cached value.
- Stored the cached values on the the outermost model `EstablishClaim`
  instead of `Appeal`. This way the caching is less invasive.
- Only cached values that are considered to be immutable for the
  `EstablishClaim`.