Welcome Gate Episode III --- Revenge of the cased database entries (#2491)
Two more fixes for the welcome gate.


We upcase the CSS ID since that's how it is stored in VACOLS
We update the redux store marking cases as viewed when the documents have loaded. This way if a user returns to the welcome gate, the new/continue buttons will be properly displayed.