Fix opening in new tab (#2426)
Opening a document in a new tab was broken. This PR fixes three things:

Middle clicking on a doc type will open the document in a new tab.
When in a new tab, we do not load the rest of the case file, and do not provide a back to file button.
There are tests to find regressions.