Adding docker-compose.yml, Dockerfile and Jenkinsfile.pr for dev and DSVA Jenkins Test Environment (#2448)
Added `docker-compose.yml` and the `Dockerfile` to build a test environment on-demand for the DSVA Jenkins Test Environment. This test environment will use the pipeline in `Jenkinsfile.pr` to execute the tests in the Docker container specified by the Dockerfile. This will not disrupt the TravisCi tests and is not meant to be used officially until the Jenkins test environment is complete. 

Instructions on how to use the docker-compose setup were added to the README.