Fix scroll while pages are loading (#2544)
Consider when the user jumps to a comment, or even scrolls to page number N. If pages 1 through N haven't been rendered yet, as they render they may change sizes. The user's scroll location remains constant, while the document may move up or down. To keep the user in the same location we now monitor the change in page sizes and scroll the window to keep the user fixed.