Fix jump to comment (#2636)
The first fix we made to properly jump to comments failed in validation for one specific case. When the screen was large enough such that pages could fit side by side when they were properly sized, but not when they were the default page size, we saw a lot of jitter as the pages loaded. Instead we now get all the sizes up front so we never have to guess at the sizes of pages.