Finish *fingers crossed* eFolder Express integration (#2853)
We're addressing three problems.

We were not passing the efolder_id to the frontend.
We were initializing the variable EFolderService based on a feature flag. However, initializers require a restart. We need our feature flag to force a change without a restart.
We needed to fix our fake authentication service to enable easier development with a local version of eFolder.