Stop extra calls to documents endpoint (#2866)
Because of a race condition we currently can call the documents endpoint several times in a row. Now we add a set of promises and setTimeouts to wait for the first request to finish and then use the result.

We also add a config flag to add to make local development against an efolder instance easier.