Remove single page with metadata (#2871)
When opening a document in a new tab we had a "new tab mode." In order to try to get this new tab mode to load quickly we put all the necessary information about the page in the URL. The id, the receipt date, and the type. This way the rails server didn't have to fetch data from VBMS to display this new tab. However, forking the path based on the existence of this data was problematic. It's resulted in a number of bugs, but currently the following three.