Fix PDF hang on eFolder integration  (#2997)
Currently when we turn on the eFolder integration, EVERYTHING HANGS when switching PDFs. After some digging we found that lodash's omit function accepts keys to omit, or paths to omit (where a path is defined with .). The idea being that you can deep omit into a nested object.

This worked fine with our old pdf urls since they were local and didn't contain the part of the URL with .. But when we move to eFolder we needed an absolute url which started with efolder. This matched the regex they use to determine if we are specifying a path, which triggered a deep clone of the object. Deep cloning this particular object takes a long time since the PDF objects are really big!