Cancel loading PDFs when unused (#3080)
This PR addresses two problems. First, when our getDocument function is called for a file that is not currently loaded, we still try to get the file. Second, when switching quickly between documents we currently destroy the worker thread of documents that we load if we realize we don't need them. With this change we will cancel the network request before we retrieve the document.