Add combine to reducers (#3095)
Currently we have one reducer. We need to split this into several organized reducers. First, we need to add the redux function combine. This creates a sub object in our state. We add that to all of our state instances. This sets us up to be able to start splitting things out into their own reducers.