Move set dimensions into pdfPage (#3177)
We're moving the setDimensions calls into the PdfPage component.

Up until this point, everything having to do with rendering the PDF has been done in Pdf.jsx. PDFJS deals with everything on a per page basis, and our PDF component renders multiple pdfs at once (the one currently in view, and upcoming pdfs to make future rendering fast). This means that all state associated with rendering the PDFs in Pdf.jsx has to be a multi-dimension array indexed by the file and page number we're interested in.

We can gain a lot of simplicity by moving as much of this logic into the newly created individual page components. Eventually these components will be in charge of rendering themselves. We start by moving the logic that determines the size of each page into the PdfPage component.

This is done in three steps:

We move the pageDimensions variable in the local state from Pdf.jsx into the Redux store.
We move the logic for calculating pageDimensions into the PdfPage component.
We replace all instances of the old pageDimensions variable with the new Redux state.
After these steps it became clear that we didn't need to store pageBounds in the Redux state since it was a duplicate of our pageDimensions state. We removed this dead code.