Move text into the comment layer (#3222)
With the creation of the comment layer we created a new div that is above the text layer div in the DOM, and is preventing click events from making it down to the text layer. While I'd prefer to keep them separate, it seems like the simplest solution is to make the CommentLayer component contain the text layer div. Then it can receive all the clicks.