Fixing jumping to comments and null pointer in analytics (#3250)
When we did a refactor to use combineReducers we moved our state to a new substate: readerReducer. We missed three references that didn't reference the sub object.

When we refactored pdf to use PdfPage, it seems like we introduced some race condition where we could calculate the sizes of all the pages before having references to all of them. Our tests didn't catch it since it occurred non-deterministically. Now we make sure we have all the page references before jumping to a page.