Adding breadcrumbs (#3344)
Adding a generic way to implement breadcrumbs in our navbar. To auto-generate breadcrumbs you need to do the following:

Make either Route or PageRoute objects the children of the NavigationBar.
Give each of those Routes/PageRoutes a breadcrumb attribute that has the name of the breadcrumb.
Order the routes so that sub pages come after higher level pages.