Intake | Add `establish_end_product` method to RampElection
This will allow us to see if the end product of a ramp election
is still active or not.

This is just a back end change for now.

Also create `Generators::EndProduct` to make it easier to mock
end products. This change includes scoping mocks to specific
veteran ids to improve test coverage.

connects #4141