Release connections when making API calls (#4163)
The core of this PR is to release any DB connections we're holding before making any API calls. This way we can reduce the number of DB connections our app needs. We also add code in this PR to create a new rails environment. This environment fakes our most dependencies, but allows you to connect to your own VACOLS copy.