Intake | Provide scafolding for different finish pages
This breaks up the finish page into two page components for the
two different form types. This is a prerequisite to begin
development on the finish page for Ramp refilings.

Also turn `getRampElectionStatus` into a more generic selector
that gets the status of the current intake.

This is purely a refactor (no new functionality) and shouldn't
need any testing beyond the existing suite of unit/feature tests.

connects #3966