Intake | Generate Ramp Issues for prior election when starting refiling
In order to check whether the contentions had been created for the previous
RAMP election EP, we need to load them from VBMS on search.

Decided to go ahead and generate the RAMPIssues from those contentions on
the search step to save costly additonal calls to VBMS for the contentions.

Testing Instructions
85555555 has been configured in demo to not have contentions mapped to the
EP for the Ramp Election. First complete the RAMP Election intake for this
file number, then attempt to complete the RAMP Refiling intake for it. You
should see the no contentions error.

Notes
The check for an EP having no contentions handles an edge case that
was not listed on the AC. The copy for the error is temporary and I'll check
in with Abby on what the final copy should be.

connects #3966