Intake | Finish refiling page without submit
Add the form for finishing RAMP Refiling. Left out the submit logic for
a future PR, but should have all the elements needed on the form.

Testing Instructions
- Start a refiling, go to the finish step and check all the pretty the boxes

Notes
Since this is a time sensitive feature, we made some simlifications to the
design to save time. We will return to the design of these pages when we
do user testing. Abby will validate the final UX of the page.

connects #3966