Adding slider for overscan count (#4273)
We want to experiment with users to determine what overscan count produces the best windowing results. This "feature" is behind a featureToggle that we will only hand out to a couple of our trusted attorney's judges. We'll ask them to play with it to find what overscan count has the best performance.