Fix failing queue_spec test
One of the specs in feature/queue/queue_spec.rb attempts to click on a link with a number in the link.
The number in that link is rendered in the frontend with .toLocaleString, which means the number is formatted with commas in most locals (e.g., 1234 -> 1,234).
The spec selector does not account for this, and so when there is a link that says "Open 1,000 documents", the spec attempts to find and click a link that says "Open 1000 documents" (notice the missing comma), and fails.

This makes the spec a little more robust, asking it to look for a delimited string.