Adds a trailing space to random whitespace (#4918)
In reader_spec, some tests rely on a random string of whitespace to test both 1. in some situations, blank input is invalid, even if the input contains arbitrary strings of whitespace, and 2. in some situations, blank input is valid and means that something should be deleted.

In CircleCI, we were seeing random failures of reader_spec for tests of the latter case (delete). We are not sure about this, but our theory is that random whitespace strings with certain endings, the browser may have been interpreting that input as an "enter key" press as opposed to a text-field value of whitespace, which would cause unexpected test behavior. We fix this here by always appending a trailing space to the random whitespace.