Moving componentDidUpdate into componentDidMount (#5494)
Resolves https://github.com/department-of-veterans-affairs/caseflow/issues/5492

### Description
We had code to scroll to the most recent location of a user on the document list page. However, this code was inside componentDidUpdate. Until this change: https://github.com/department-of-veterans-affairs/caseflow/commit/c9ec40973bb745a93a933ff45e7f4d8dacd91253, we erroneously called componentDidUpdate very frequently because the state was always changing with coordinates for the dropdowns. Now we rarely call it.

The solution is to move the scrolling code into componentDidMount. It's where it always belonged. We only need to scroll when we return to this page and it mounts.
![documentlistscrolls](https://user-images.githubusercontent.com/3885236/39890711-99a6d8ee-5469-11e8-81eb-1fe085bf48d5.gif)

### Acceptance Criteria 
- [ ] Returning to the document list from a document scrolls to the document the user just visited.