Make CaseTitle self contained (#5845)
As we start using CaseTitle in more places in the code it becomes increasingly clear that we pass the same components into CaseTitle. This PR moves those elements inside the CaseTitle component (and create a new CaseTitleScaffolding component) so we only have to pass the bare minimum set of elements to the component for it to display correctly.