Add case snapshot to case details page (#5837)
Connects #5744.

Moves some information about the appeal to a highlighted box at the top of the page that we are calling Case Snapshot.

| Before | After |
| ------ | ----- |
| ![image](https://user-images.githubusercontent.com/32683958/41128083-157a4c7c-6a7b-11e8-9e9d-cbd9be95ddef.png) | ![image](https://user-images.githubusercontent.com/32683958/41127981-a700c5e6-6a7a-11e8-8c55-65c630272fd6.png) |

Right now the application and presentation code is very intertwined and I am definitely open to suggestions for how we could untangle the two. I took a few stabs at doing just that but couldn't come up with a solution that I particularly liked. If we can't think of anything better here then I think it is fine to wait to separate this component out into app/presentation pieces until there are other use cases when the improvements may become more clear.