Do not make document count requests for paper cases (#5865)
When [working on ignoring errors](https://github.com/department-of-veterans-affairs/caseflow/pull/5863) related `ClientRequestError`s that are thrown because there are no documents in VBMS for a given Veteran, it occurred to us that we should not make any requests to the `document_count` endpoint if the appeal is a paper case because there will be no documents for that case. This PR makes that change.