Ignore ClientRequestError when hitting document_count endpoint (#5863)
Similar to #5830. Example of an instance where #5832 would be beneficial as well.

This change 1) prevents Sentry from sending an alert to slack and 2) returns a 400 response instead of a 500. We are hitting this error because we make a request to the `/appeals/{vacols_id}/document_count` endpoint for an appeal and BGS (by way of efolder) cannot find the Veteran associated with that appeal.

Confirmed this is a legitimate response by bypassing this BGS check and requesting the documents for this Veteran's file number directly from VBMS (v1 and v2 endpoints) and the file number does not exist in VBMS either. Perhaps this indicates that the case is a paper case? (cc @dannysellers)

Link to [the slack discussion](https://dsva.slack.com/archives/C6E41RE92/p1528720066000336) with additional thoughts on the approach.