Stop loading every appeal for a second time. (#6002)
### Description
We are currently calling `find_or_create_by_vacols_id` for every appeal. However, `find_or_create_by_vacols_id` is overridden and calls `check_and_load_vacols_data!` Which makes a VACOLS call to load the data. This PR passes appeals in that we've already queried to reduce the number of calls we make.

### Acceptance Criteria 
- [ ] Each appeal should only be loaded once at the beginning of the query.

### Testing Plan
1. Go to my queue locally and ensure we do not reload data for every appeal.