Allow AMA intakes when Veteran has no rated issues (#6390)
Connects #5868

This allows a Veteran without Ratings to still go through the intake process.  If they have no ratings, a message will be displayed to the user notifying them of this in case they need to add ratings.  Otherwise they can proceed with non-rated issues only.

Note: this does not create non-rated End Products for the non-rated issues yet.