First pass at creating a way for users to enable/disable global feature toggle (#6455)
This creates a way for us to have our team members enable and disable global feature toggles using the demo UI. 

<img width="524" alt="image" src="https://user-images.githubusercontent.com/31223104/43529852-2d8e3f88-957a-11e8-84ab-5bf476187e57.png">

This can be viewed on /test/users page. 

I ended up doing a reload after the POSt instead of updating the state in the app.  The interaction works _ok_, but let me know if it makes sense to update the app state instead.