Highlight current assignee in case search if it is the logged-in user (#6559)
Connects #5179.

Displays the green text "Assigned to you" in the "Assigned to" column of the case search results page if the case is assigned to the currently logged-in user. Displays the location code of the case if it is not assigned to the logged-in user. Prior to this change the "Assigned to" column always displayed the location code, even if the case was assigned to the current user.

| Assigned to you | Assigned to somebody else |
| --- | --- |
| ![image](https://user-images.githubusercontent.com/32683958/43906950-764a6f50-9bc2-11e8-84e9-512020065c00.png) | ![image](https://user-images.githubusercontent.com/32683958/43906882-51d823ec-9bc2-11e8-878f-6686a26f6c64.png) |