Fixing appeal details (#6544)
Connects #6540

### Description
In the last refactor we created an `appeals` and `appealDetails` section of the state. However the `appeals` section which contains only basic appeal information was structured without the `attributes` sub-object and the sub keys were in camel case. The `appealDetails` did have the `attributes` sub-object and the sub keys were still snake cased. Finally, there was overlap between the `appealDetails` and `appeals` section. This removes that overlap and creates consistency in how we represent our state.

We've created two new selectors: `appealsWithDetailsSelector` and `appealWithDetailSelector` to get the combined basic and detailed appeal information.

### Testing Plan
1. All the tests still pass and a manual walkthrough shows no changes.