Redirect VSO users to organizational queue (#6660)
This PR makes it so requests to `/queue` from VSO users will redirect them to the URL of the organizational queue for the VSO they are associated with. This enables us to point all VSO users to the `/queue` endpoint in training so we don't have to remember the URLs for each VSO. Additionally, this prevents VSO users from accessing personalized task queues until (if?) those are enabled.