Display Veteran date of birth correctly for BEAAM appeals (#6671)
Connects #6550.

When we display the Veteran's date of birth in case details, we use the `veteranDateOfBirth` field from the appeal which comes from the `veteran_date_of_birth` field in the appeal serializers. For `LegacyAppeals` the `veteran_date_of_birth` field comes from VACOLS as a datetime object (I think), and was serialized in the `legacy_appeal_serializer` in the format "1989-11-23". The `veteran_date_of_birth` for BEAAM `Appeals` comes from the `Veteran` object's call to BGS for that value where it is returned in the format "11/23/1989", which is the value the `appeal_serializer` passed to the frontend. On the frontend the `DateString` component assumed the `veteranDateOfBirth` was always in the format "1989-11-23", and made a best guess at what the value should be for the values in the format "11/23/1989"... that best guess was never correct and resulted in incorrect birthdates.

This PR changes 1) the `DateString` component to expect Veteran birthdates in the format "11/23/1989", and 2) the `LegacyAppeal` serializer to return birthdates to match the format returned by BGS ("11/23/1989").