Create Task methods to enable generic task completion (#6677)
Connects #6678.

We need to support each of the three generic task actions (`Mark complete`, `Assign to team`, `Assign to individual`--I foresee `Assign to self` just being a shortcut on the frontend for "assign to individual" with the current user's ID in the assignee parameter) in the `Task` model. This is a first pass at doing that with the `Mark complete` task action.

I still need to write tests, but I'd like a second set of eyes validating this approach given our tech spec.