Filters documents array by doc id, corrects "written_by" attribute (#6694)
Connects to https://github.com/department-of-veterans-affairs/caseflow/issues/6645

Previously, we were returning DECASS records without document ids, we should filter those out. Also, we were returning the CSS ID of the person who the appeal was currently assigned to, not the attorney who wrote the doc.

We also only need the CSS ID of the user who assigned the current task, not for all documents.