Check VSO has access to appeal on search screen (#6790)
Resolves #

### Description
When a VSO searches for a case, we check all the appeals the veteran has and only surface those the VSO is a power of attorney on.

This means we
1) Find all the claimants associated with all the appeals associated with the Veteran
1) Get the claimants' participant ids and query BGS to get their POAs
1) Check the POAs against the POAs the user belongs to
1) For each appeal that has the POA as a representative, show it on the search screen 

Most of the code is fake data setup and tests. I think this is an important scenario to get right. We don't want to leak Veteran data to POAs that shouldn't have access.

We also rely on a new BGS integration where we can pass an array of participant ids get get all the corresponding VSOs.

### Testing Plan
1. Run `rake db:seed`
1. Sign in as the VSO user
1. Go to `/organizations/american-legion`
1. Search for the veteran number: `701305078`
1. Only one result should appear
1. Do the same as BVASCASPER and two results should appear