Determine AOD by grants or age (#6836)
Resolves #6274

### Description
We need to change how we determine AOD. Currently we have a column on the Appeal table that marks the case as AOD or not. We've determined this is a bad structure for a few reasons:
1) AOD is attached to a claimant not an appeal
1) AOD needs to be automatically marked when a claimant is over 75

In order to achieve this we're removing the AOD field on the appeal table and creating a new advance_on_docket_grants table which will have a one-to-many relationship with claimants. This new table will contain the reason a case is being marked AOD.

In addition we're adding a date_of_birth field to the claimants table so we can determine when a case should be marked AOD due to age. There's a chance we'll need to build a UI in Intake for this in case the DOB field on the NOD is the claimants DOB. More to come there.

Note I know the DB migration is not backwards compatible with the application code around BEAAM appeals. But I think it's okay since it should only impact BEAAM appeals which only have a very small number of users. There will be a 5-10 minute span where anyone who requests BEAAM appeals might get an error. I can split this up if we deem that risk to high.

### Testing Plan
1. The automated tests exercise the feature.