Removing migration (#7033)
### Description
When I merged the `201809112051441_index_claimants_on_date_of_birth.rb` migration I accidentally put an extra digit in the timestamp 🤦‍♂️ . This messes up the schema file, as this migration will always have the largest timestamp and the schema file will always use it. In order to fix it, we need to rename this file, but we can't do that since the name is already stored in Rails' schema migrations table. I think the easiest way to fix this is to rollback this particular change with `rake db:rollback VERSION=201809112051441` on prod, uat, preprod, and demo. And asking all developers to run it as well. Then I can remove the file in this PR and deploy. We're going to rearchitect this column a little bit anyway--so instead of recreating it, I'm just going to delete it. It's not currently in use.

### Acceptance Criteria
- [ ] Schema.rb stores the timestamp of the most recent migration, not always `201809112051441`

### Testing Plan
1. Run `rake db:rollback VERSION=201809112051441` and ensure the schema file is updated properly.