Clicking Caseflow logo in Reader links to Queue (#7077)
Connects https://github.com/department-of-veterans-affairs/appeals-support/issues/3067#issuecomment-423214771.

This change was made in #6967 but overwritten by a related change in #6932, so it worked as expected from Friday until Tuesday morning. When we released on Tuesday the code was changed so that the logo pointed to the base path. This PR fixes that.