Add tasks/:id/assignable_... endpoints (#7183)
Connects #6703. Adds endpoints for tasks to call to get the list of people or organizations that the task can be assigned to.

## To the reviewer
The meat of the review starts with the [new actions in the `TasksController`](https://github.com/department-of-veterans-affairs/caseflow/pull/7183/files#diff-f4416012d567d02da854cfbc9e2df2caR98).