fixing generate_profile_date (#7193)
`generate_profile_date` was changed from a date => datetime as part of https://github.com/department-of-veterans-affairs/caseflow/pull/7139

This caused the generator to generate ratings on the same day since the lazy eval of the date list was different by a 1/10th of a second each time. This broke the UI which only displayed one set of ratings per day.