Make generic task dropdown (#7175)
Connects #7102 

### Description
First PR to fix the task dropdown. This change adds a new task drop down that will be used for all cases. We replace the attorney dropdown with this new one, and set up the structures to add new options.

We also move checkout flows to have unique URL paths to move state from Redux into the URL about which flow we're in.

### Testing Plan
1. Manually test various checkout flows.