canceling intake from add_issues should redirect (#7353)
Connects #7312 

### Description
The current Finish page will redirect back to intake start (`/intake`) if the formType is undefined, which is cleared when canceling an intake. The new AddIssues page should be consistent with that behavior.

I'm not really sold on using a falsy `formType` to redirect since this is navigation by side effect, but this is simple for now. I think once we switch over to the new AddIssues page by default we can edit the `submitCancel` action to do the redirection.

![demo](https://user-images.githubusercontent.com/279406/46788827-df46a880-ccef-11e8-8986-5cf2576b6171.gif)

### Acceptance Criteria
- [x] Canceling any intake from the AddIssues page will redirect back to `/intake` as the current Finish page does

### Testing Plan
1. Start any intake, get to the review screen
2. Visit /intake/add_issues
3. Cancel the intake and verify that you are redirected back to /intake (currently there is a blank screen)