Return all tasks and change indexing on FE tasks (#7365)
Connects #7102

### Description
We split this out from https://github.com/department-of-veterans-affairs/caseflow/pull/7214.

This PR fixes a bug where we only return a subset of tasks from the task_controller's for_appeal method. However, once we start returning all tasks we need to change how we index them. This moves us to use the externalAppealId to index legacy tasks and the task id to index modern tasks.

### Testing Plan
1. Test Co-located flow for AMA and legacy tasks
1. Test attorney to judge flow for AMA and legacy tasks