Judge assign modal (#7214)
Connects #7102

### Description
Move the assign widget to a modal on the case details page. Also noticed we were only passing one task to the case details page. We want to return all tasks.

Note: Assigning the task to an `other` attorney, and then trying to assign to a different `other` attorney fails because the assign/reassign logic doesn't work in that case. But I think this is probably broken on master as well.

### Testing Plan
1. Login as a judge
1. Go to the assign queue
1. Click on an appeal
1. Select the task `assign to user`
1. Assign to a user in the modal