Assign Schedule | Ro not assigned hearings (#7376)
Resolves #7106

### Description
Please explain the changes you made here.

### Acceptance Criteria
**1.) There are no veterans to assign.** 

Assuming we are pulling Veterans for that regional office from location 57. So, it makes sense to have RO specific no veterans messages because some RO's may still have veterans to assign. Is this right? In that case we can do this to perserve the work that was done already. In the future a smaller messaging might be better: 

The only difference is that there is no gray background because we are no longer disabling the days filter. 

_Copy_
Title: There are no schedulable veterans 
Body: Please verify that this RO has veterans to assign a hearing


### Testing Plan

1. Go to assign hearings and don't assign Ro hearings. Show error
<img width="1215" alt="screen shot 2018-10-15 at 9 42 49 am" src="https://user-images.githubusercontent.com/23080951/46958109-0c37ea00-d067-11e8-8b17-e7abb2665a70.png">